minGRU kernel: consume fp16/bf16 directly instead of casting at the boundary
Measured on an H100 at 131k tracks/batch: fp16 GEMMs are worth +58 % over
strict fp32 (1.69 -> 2.67 M tracks/s), but stacking fp16 autocast on the TF32
deployment path made it SLOWER (3.75 -> 2.90 M) because the encoder had to
materialise an fp32 copy of the (T, 4H) projection at the scan boundary.

The kernel now converts on load ( on a load that happens
anyway = a register convert, free) and lets the output dtype follow the input,
so a half-precision encoder halves the global traffic of the largest
intermediate.  The recurrence still accumulates in fp32 inside the kernel: a
product of up to MAXL gates in fp16 would lose mantissa and can underflow the
6e-8 subnormal floor.

bench_infer_flat gains --encoder-dtype {float32,float16,bfloat16} (the
encoder_autocast_dtype hook, so normalisation, Fourier, heads and the fp64 seed
stay outside it).  Tests: fp16/bf16 parity against the fp32 scan and an
explicit check that the fp32 path is unchanged.

+++ src/track_regression/mingru.py
@@ -310,13 +310,12 @@ class MinGRUCLSEncoder(nn.Module):
             h = x[0]
             H = self.hidden
             for layer in self.layers:
-                # The scan runs in strict fp32 even under a reduced-precision
-                # autocast: it is 0.25 % of the FLOPs, but a product of up to
-                # 20 gates in fp16 would lose mantissa and can underflow the
-                # 6e-8 subnormal floor.  The projections (all of the cost) stay
-                # in whatever dtype autocast chose.
+                # No cast at the boundary: the kernel takes fp32/fp16/bf16 and
+                # converts on load in registers, accumulating the recurrence in
+                # fp32 regardless.  Materialising an fp32 copy of this (T, 4H)
+                # tensor instead costs ~20 % of the forward (measured).
                 h = mingru_bidi_packed(
-                    layer.in_proj(h).float().contiguous(), cu, H, self.max_len)
+                    layer.in_proj(h).contiguous(), cu, H, self.max_len)
             term = torch.cat(
                 [h[cu[1:].long() - 1, :H], h[cu[:-1].long(), H:]], dim=-1)
             pooled = self.pool_proj(self.pool_norm(term))
